requirements.txt
boto3==1.35.71
Pillow==11.0.0
protobuf==5.29.1
python-dotenv==1.0.1
Requests==2.32.3
